requirements.txt
requests~=2.28.2
gradio_client~=2.0.0
huggingface-hub~=0.36.0
Pillow~=10.1.0
opencv-python~=4.8.1.78
numpy~=1.26.4
torch~=2.6.0+cu124
diffusers~=0.35.2
transformers~=4.57.3
mmcv~=2.2.0
deep-translator~=1.11.4
scipy~=1.15.3
trimesh~=4.10.0
svgelements~=1.9.6
shapely~=2.0.6
nltk~=3.9.1
spacy~=3.8.2
pandas~=2.2.3
torchvision~=0.21.0+cu124
pdfplumber~=0.11.0
PyPDF2~=3.0.1
mmdet~=3.3.0
gpt4all~=2.8.2
langchain~=0.3.26
whisper~=1.1.10
openai-whisper~=20231117
torchaudio~=2.6.0
soundfile~=0.13.1
sounddevice~=0.5.1
librosa~=0.11.0
onnxruntime~=1.22.1